+++ swiftide-core/src/query_traits.rs
