+++ swiftide-core/src/search_strategy.rs
@@ -1,0 +1,6 @@
+use crate::querying;
+
+#[derive(Debug, Default, Clone, Copy)]
+pub struct SimilaritySingleEmbedding {}
+
+impl querying::SearchStrategy for SimilaritySingleEmbedding {}
