+++ swiftide-integrations/src/qdrant/persist.rs
@@ -3,14 +3,12 @@
 //! This integration allows the Swiftide project to use Qdrant as a storage backend.
 
 use std::collections::HashSet;
-
-use anyhow::Result;
-use async_trait::async_trait;
-use qdrant_client::qdrant::UpsertPointsBuilder;
-
 use swiftide_core::{
-    indexing_stream::IndexingStream, node::EmbeddedField, node::Node, traits::Persist,
+    indexing::{EmbeddedField, IndexingStream, Node, Persist},
+    prelude::*,
 };
+
+use qdrant_client::qdrant::UpsertPointsBuilder;
 
 use super::{NodeWithVectors, Qdrant};
 
@@ -54,7 +52,7 @@ impl Persist for Qdrant {
     ///
     /// This function will return an error if the node conversion or storage operation fails.
     #[tracing::instrument(skip_all, err, name = "storage.qdrant.store")]
-    async fn store(&self, node: swiftide_core::node::Node) -> Result<Node> {
+    async fn store(&self, node: Node) -> Result<Node> {
         let node_with_vectors = NodeWithVectors::new(node.clone(), self.vector_fields());
         let point = node_with_vectors.try_into()?;
 
@@ -83,7 +81,7 @@ impl Persist for Qdrant {
     ///
     /// This function will return an error if any node conversion or storage operation fails.
     #[tracing::instrument(skip_all, name = "storage.qdrant.batch_store")]
-    async fn batch_store(&self, nodes: Vec<swiftide_core::node::Node>) -> IndexingStream {
+    async fn batch_store(&self, nodes: Vec<Node>) -> IndexingStream {
         let points = nodes
             .iter()
             .map(|node| NodeWithVectors::new(node.clone(), self.vector_fields()))
