+++ swiftide-integrations/src/qdrant/retrieve.rs
@@ -1,0 +1,3 @@
+use swiftide_core::Retrieve;
+
+impl Retrieve
