+++ swiftide-integrations/src/scraping/html_to_markdown_transformer.rs
@@ -3,7 +3,7 @@ use async_trait::async_trait;
 use derive_builder::Builder;
 use htmd::HtmlToMarkdown;
 
-use swiftide_core::{node::Node, Transformer};
+use swiftide_core::{indexing::Node, Transformer};
 
 #[derive(Builder)]
 #[builder(pattern = "owned")]
