+++ swiftide-query/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod query;
 pub mod query_transformers;
 pub mod retrievers;
-pub mod search_strategy;
-pub mod traits;
