+++ swiftide-query/src/query/mod.rs
@@ -1,9 +1,3 @@
-#![allow(clippy::module_inception)]
-
 mod pipeline;
-mod query;
-mod query_stream;
 
 pub use pipeline::Pipeline;
-pub use query::*;
-pub use query_stream::QueryStream;
