--- swiftide-query/src/search_strategy.rs
@@ -1,6 +1,0 @@
-use crate::traits;
-
-#[derive(Debug, Default, Clone, Copy)]
-pub struct SimilaritySingleEmbedding {}
-
-impl traits::SearchStrategy for SimilaritySingleEmbedding {}
